refactor(core): move streaming codec tests to tests/unit/codec/
Match the project's existing test layout: every other codec file in
crates/core/src/codec/ pulls its tests from
crates/core/tests/unit/codec/<name>_tests.rs via #[path]. The streaming
codec tests were inlined in src/codec/streaming.rs in the initial
commit; move them out to follow the same convention so source files
stay slim and all codec test files cluster under tests/unit/codec/.

No behavior change. 283 core lib tests still passing.

Refs NMF-90.
